refactor(Link): move to typescript
